requirements.txt
openai
pandas
requests
beautifulsoup4
python-dotenv
streamlit
tqdm
pyperclip
